Cleaned up learners.md
For future classes, there needs to be a new learners(month-year).md file created, and you should put each group TF, and the names of the students. This allows students to come in, and just change the line their name is on with their github information. So PR's should be much easier in the future, and can just be accepted, without conflicts.